+++ packages/vue-dot/src/patterns/FilePreview/locales.ts
@@ -1,3 +1,0 @@
-export const locales = {
-	previewNotAvailable: 'Impossible de prévisualiser le fichier'
-};