+++ CHANGELOG.md
@@ -2,14 +2,14 @@
 
 ### Vue Dot
 
+- ✨ **Nouvelles fonctionnalités**
+  - **UploadWorkflow:** Ajout de la prévisualisation des fichiers avant validation ([#2336](https://github.com/assurance-maladie-digital/design-system/pull/2336))
+
 - 🐛 **Corrections de bugs**
   - **HeaderBar:** Correction des options de personnalisation ([#2314](https://github.com/assurance-maladie-digital/design-system/pull/2314)) ([aa4696b](https://github.com/assurance-maladie-digital/design-system/commit/aa4696bdb47cb8e4fa9630d6002893b0a6543fe3))
 
 - ✅ **Tests**
   - **testUtils:** Ajout d'un test sur la fonction `html` ([#2315](https://github.com/assurance-maladie-digital/design-system/pull/2315))
-
-- ✨ **Nouvelles fonctionnalités**
-  - **UploadWorkflow:** Ajout de la possibilité de prévisualiser les fichiers avant validation ([#2336](https://github.com/assurance-maladie-digital/design-system/pull/2336))
 
 ### ESLint Config